Prioritize raw interaction dispatch
In the event that we're dealing with raw interactions, we want to
consume that interaction as soon as we reasonably can.

Currently it means waiting for other resolutions to occur which
ultimately don't matter outside of library specific interaction
constructs.

Rather than waiting for said resolutions, let's dispatch the raw
interaction as soon as we are able.